Move old nitrite.go to internal
