+++ internal/result.go
@@ -1,0 +1,31 @@
+package internal
+
+import (
+	"crypto/x509"
+)
+
+// Result is a successful verification result of an attestation payload.
+type Result struct {
+	// Document contains the attestation document.
+	Document *Document `json:"document,omitempty"`
+
+	// Certificates contains all of the certificates except the root.
+	Certificates []*x509.Certificate `json:"certificates,omitempty"`
+
+	// Protected section from the COSE Sign1 payload.
+	Protected []byte `json:"protected,omitempty"`
+	// Unprotected section from the COSE Sign1 payload.
+	Unprotected []byte `json:"unprotected,omitempty"`
+	// Payload section from the COSE Sign1 payload.
+	Payload []byte `json:"payload,omitempty"`
+	// Signature section from the COSE Sign1 payload.
+	Signature []byte `json:"signature,omitempty"`
+
+	// SignatureOK designates if the signature was OK (but certificate could be
+	// invalid, not trusted, expired, etc.)
+	SignatureOK bool `json:"signature_ok"`
+
+	// COSESign1 contains the COSE Signature Structure which was used to
+	// calculate the `Signature`.
+	COSESign1 []byte `json:"cose_sign1,omitempty"`
+}
